[pipes] make `PipesClient.run` synchronously return `PipesClientInvocationResult` (#16898)
## Summary & Motivation

Change `PipesClient.run` to synchronously execute a pipes process and
return a new `PipesInvocationResult` object instead of acting as a
generator. Results are available
via`PipesInvocationResult.get_results()`.

An error is thrown if results are expected but not returned. This fixes
the problem with `PipesClient.run` where, if you did not `yield from`
it, it would silently not execute with no error because it was a
generator function.

## How I Tested These Changes

New unit test.