feat(dbt): add `DagsterDbtTranslatorSettings` (#16889)
## Summary & Motivation
As discussed in https://github.com/dagster-io/dagster/pull/16852,
enabling asset checks is an all-or-nothing operation. In order to not
confuse users about this behavior, we will only allow users to enable
this feature through a binary toggle, rather than setting the metadata
on individual dbt tests.

Here, we use `DagsterDbtTranslator` as the object that contains the
settings that a user can configure to enable features on their dbt
project when consuming it with Dagster.

One thing to note here is that custom subclasses that implement their
own `__init__` will need to also pass down the settings object if they
are opting into asset checks.

## How I Tested These Changes
pytest