Add a query for querying asset evaluations by evaluation ID
Summary:
This will power the table view where we show what happened to each asset for each evaluation Id. Left out a cursor for now because this query has an index and is limited to the size of the asset graph, but I think to have efficient limiting and cursoring we would need an index by both evaluation ID + id.

Test Plan: BK