Rename get_pipes_env_vars to get_bootstrap_env_vars. Refactor to call generalized get_bootstrap_params (#16939)
## Summary & Motivation

This is both a refactor and a change in terminology. I'm in the process of writing the launch post on Pipes, and describing the initial parameter process as "bootstrapping" has been very helpful.

Furthermore, I also want to make it clear that while environment variables are the most common way to pass bootstrapping parameters, it is not the only way. The decision about how to bootstrap is the responsibility of a particular context injector, not a hardcoded behavior in the system.

The code level is feels good to separate the encoding process from setting up the logical parameter dictionary as well.

## How I Tested These Changes