[Asset-graph] Make cache key deterministic (#16947)
## Summary & Motivation
Noticed that the cache key we use isn't created in a deterministic way,
not sure exactly why (specualation: could be the order from the database
for the assetKeys from the workspace query isn't deterministic). Let's
make it deterministic by sorting the keys.

## How I Tested These Changes

locally reloaded the page a few times and made sure we always use the
cache. Also restarted the webserver