Replace Next.js' static `assetPrefix` with `__webpack_public_path__` to support dynamic path prefix (#16967)
## Summary & Motivation
This pr addresses a performance issue reported by a user:
https://dagster.slack.com/archives/C01U954MEER/p1695936650103649

This is an alternative approach to
https://github.com/dagster-io/dagster/pull/16941 that avoids needing to
string replace javascript files at run time and instead replaces at
buildtime using `__webpack_public_path__`.

This solution is pretty hacky for webworkers and may need tweaking in
the future if we start SSRing more of the app. For workers we grab the
path from the global location object.

## How I Tested These Changes

Built the app and ran `dagster-webserver -p 4444
--path-prefix=/test1235` and made sure the app still worked including
webworkers.

This doesn't affect cloud or local development (I tested local
development and it just ignores __webpack_public_path__).