Disable partitionStatusCounts resolver for asset backfills (#16985)
This PR disables the `partitionStatusCounts` resolver for asset
backfills, returning a null value when the targeted backfill is an asset
backfill.

This resolver assumes a unique run per partition, and returns the run
status per partition to display in a progress bar. However, for asset
backfills, this frequently doesn't apply:
- Asset backfills can kick off multiple runs for the same partition key
(i.e. a weekly partitions def downstream of a daily partitions def)
- Asset backfills can target different partitions per asset, which
doesn't display well in a condensed progress bar (and currently raises
an error)

We'll need to rethink how asset backfill statuses should be displayed,
but early thinking is displaying counts by status. For asset backfills
we can rely on the `assetBackfillData` resolver for this information.