Fix negative asset backfill counts caused by reexecution (#16991)
Users reported seeing negative asset backfill counts, which was caused
by:
- Kicking off backfill runs that failed
- Before the backfill completed, they successfully manually reexecuted
the failed runs. This caused the re-executed partitions to be considered
as both materialized and failed.

This PR amends the counts to remove partitions from the "failed" counts
if they are also materialized.

There can be an edge case where a partition originally succeeds, but
fails when manually re-executed. In this case, it will display as
successful, but I think this is expected behavior.