required_assets_and_checks_by_key on InternalAssetGraph (#16924)
Rework of https://github.com/dagster-io/dagster/pull/16711

We want a `canExecuteIndividually` resolver on asset checks, in order to
disable the 'execute' button if it's just going to error. We'll support
that with `asset_graph.get_required_asset_and_check_keys(check_key) >
0`.

The next diff will add atomic_execution_units to ExternalAssetCheck, and
implement this on ExternalAssetGraph. We don't actually need this
implementation on InternalAssetGraph currently, but I think it's good
for consistency?