Add requestedAssetMaterializations to the tick graphql type
Summary:
This is much simpler to query for when rendering the tick timeline and allows us to avoid a separate query to the evaluation table - no need to dig into specific rule evaluations or anything like that. (If we did want to show skips and discards, then we would need the rule output).

Added a new graphql type with an asset key and >=0 partition keys. If it's 0, it means the asset is unpartitioned (it being in the list means that it had to have something requested, so there's no ambiguous states there). If it's >=1, the specific partition keys are included on the object.

Test Plan: BK