Add get_materialize_result to PipesClientCompletedInvocation and PipesSession (#17020)
## Summary & Motivation

This adds `get_materialize_result` to `PipesClientCompletedInvocation`
and `PipesSession`. For the common case that a pipes invocation creates
a single asset, this is a simpler API that returns a single
`MaterializeResult`. Then type signature of `@asset` can be simpler,
just `MaterializeResult`. (in practice almost no one correctly annotates
the streaming version of `@asset` at all)

I think we should with this API in all examples.

## How I Tested These Changes

BK