Wrap observable source assets in AssetsDefinition (#16620)
## Summary & Motivation

This makes an observable source asset wrappable in an AssetsDefinition.
This is an imperfect intermediate state, as as spurious asset
materialization is produced as a result of execution (verified in test
case but fixed in https://github.com/dagster-io/dagster/pull/16621) but
this verifies that the core execution machinery works.

Makes the following code possible:

```python
@observable_source_asset
def an_observable_source_asset() -> DataVersion:
    return DataVersion("foo")

defs = Definitions(assets=[
    create_external_asset_from_source_asset(an_observable_source_asset)
])
```

This `@observable_source_asset` now plugs into the rest of our
infrastructure:

1) You can execute it from `dagster-webserver` using the "Materialize"
button, which will be renamed to "Execute".
2) You can schedule this from schedules and sensors.

AMP support has to wait because auto observation is hard-coded in the
AMP logic.

## How I Tested These Changes

BK