Add support for ECS External launch type  (#16715)
## Summary & Motivation
We are using Dagster in AWS using the ECS launcher. For certain tasks we
are looking to move to Dagster the software has to be run on machines
that are in our data center. We would like to be able to use the Dagster
with ECS using ECS Anywhere, but found when trying to launch tasks with
the External launch type Dagster would fail to start tasks (attached
error in screenshot). The error was created during the RunTask boto3
call since the awsvpc networkMode is only valid if the launch type is
"EC2" or "FARGATE". This change adds a check for the launch type before
applying the aws vpc network configuration settings.

## How I Tested These Changes
I changed the dagster-aws module to be pip installed from my git
repository and was able to run tasks on External instances.

<img width="1218" alt="networkMode_error"
src="https://github.com/dagster-io/dagster/assets/5329455/ba2e5de6-519c-4941-a7b1-d48b6e3404e4">

---------

Co-authored-by: gibsondan <daniel@elementl.com>