[gql] Add user tag to reported runless event (#17042)
This PR tags runless events reported via graphQL with a tag
corresponding to the cloud user. This enables displaying which user
reported a specific event.

Prior, the user tag was only used on runs. This PR is the first place
where we will begin storing the user tag on events.

Test added in internal: https://github.com/dagster-io/internal/pull/7122