[RFC] Context to allow cloud to sort filter values (#17084)
## Summary & Motivation

This is a pretty generic way for cloud to sort filter values. It allows
Cloud to specify a sorting function for a filter by it's name.
This means that multiple filters with the same name could be targeted by
cloud and this is intentional. The thinking behind it is that you might
have a "User" filter in many parts of the app and we would very likely
want cloud to target all of them instead of just a specific one.