[ui] Support for manually reporting materialization events from asset details (#17091)
Design:
https://www.figma.com/file/V1V87q5hW7Z6G8T3T1XbTn/Manual_AddEvent?type=design&node-id=12-30&mode=design&t=oNRAFHi3u6HhLuMo-0


## Summary & Motivation


This PR adds a new option to the materialize dropdown on the asset
details page. (and ONLY this copy of the button)


![image](https://github.com/dagster-io/dagster/assets/1037212/e22e3b05-a557-4c73-b431-d3172919e419)

The new option opens a modal which allows you to manually record that an
asset materialization has taken place. If the asset is partitioned, you
can choose partitions from this modal and we pass the partition keys
through to the mutation.


![image](https://github.com/dagster-io/dagster/assets/1037212/6b5b169c-9c53-414f-9d12-b274026f4e11)

![image](https://github.com/dagster-io/dagster/assets/1037212/c792365b-71a0-4f7b-a5b3-35bb6750c2a6)


Events created in this manner seem to have a runId = "", and I'm using
that to display a tag that says "Reported" in two places:


![image](https://github.com/dagster-io/dagster/assets/1037212/969bff5d-f8f4-4562-83d4-b3eb3a8b74e5)


## How I Tested These Changes

- Just manual testing so far with partitioned and unpartitioned assets.
Everything seems to work nicely once you have these in the system, the
statuses of the partitions update as expected.

- Will add a few jest tests for this but it sounds like we want to ship
this ASAP so I may apply everyone's feedback first.

---------

Co-authored-by: bengotow <bgotow@elementl.com>