Increase live data poll rate to 2 seconds in OSS only (#17000)
## Summary & Motivation

Due to our recent changes to how we fetch asset data we can no longer
provide a query refresh countdown. This makes the app feel less alive so
for OSS only we want to increase the poll rate for data to every 2
seconds. In practice this means assets could be fetched at best every 2
seconds but it could take longer due to the chunking/queueing behavior
of AssetLiveDataProvider (which is good).

@schrockn Asked for this specifically.

## How I Tested These Changes

Locally I checked the network tab and made sure we were polling every 2
seconds