Include run worker failure reason in system logs, even if we don't include it in structured event logs (#17160)
Summary:
sometimes we don't include the reason an ECS task failed in the event
logs for the run, but it's still useful to know why it failed in the
operational logs for the agent/daemon/etc. that launched the run.

Test Plan:
Existing BK verifies that run worker monitoring is still working in
general with the correct messages
Check logs for our existing ECS cluster fleet for this message

## Summary & Motivation

## How I Tested These Changes

(cherry picked from commit b6111741223c4cdc37a2eb9684c9b6f9b5cedb99)