[2/n][pythonic config] Update config conversion logic to rely on annotation rather than Pydantic model type (#16435)
## Summary

PR stack which re-implements #16257, in hopes of making it easier to review.

This PR holds the meat of the actual changes. It reworks some of the underlying conversion code to look at the annotation (e.g. actual Python type objects) on config classes instead of the Pydantic internal type representations, when converting to Dagster config types under the hood.

## Test Plan

Existing unit tests.