fix(dbt): skip copying the partial parse file if it exists in the target (#17214)
## Summary & Motivation
As the title.

From https://docs.python.org/3/library/shutil.html#shutil.copy:

> Copies the file src to the file or directory dst. src and dst should
be path-like objects or strings. If dst specifies a directory, the file
will be copied into dst using the base filename from src. If dst
specifies a file that already exists, it will be replaced. Returns the
path to the newly created file.

But in windows, this information is false, as an error pops up if the
file at the destination already exists. In that case, skip the copy.

## How I Tested These Changes
pytest, add assertion that partial parse file is not overwritten if it
already exists