[graphql] Sort nodes in resolve_assetNodes so they are stable across reload (#17230)
## Summary & Motivation

Hey folks - This PR sorts the asset nodes returned from
resolve_assetNodes. We believe that instability in the sort order is
causing cache invalidation in a few places in the JS and also (maybe?)
causing the asset graph layout to change.

I've only observed sort differences across "reload definitions" calls,
so we could potentially move this sorting to the load step also.

I put this at the very end of the resolver after the ids have been
generated for each node so the sorter should be a simple property
retrieval. If this is slow, we could potentially dig deeper and see if
it's the sort order of the repositories that is changing and not of
individual assets, but sorting here seems like the best way to make sure
the behavior doesn't change in the future.

## How I Tested These Changes

Ran existing tests, updated them as necessary

Co-authored-by: bengotow <bgotow@elementl.com>