[ui] "Add tag to filter" on "Launched by" column tag (#17270)
## Summary & Motivation

Resolves #16835.

Show tag actions on the "Launched by" tags on the Runs table. Hover on a
tag to see an affordance to add that key/value to the filter.

I also made a slight adjustment to `UserDisplay`, to make the subway dot
letter slightly bigger and tweak the spacing to align better with tag
icons.

## How I Tested These Changes

View Runs table, verify that I can filter by the "Launched by" tag, for
all of the specified tag types.