Make adding asset evaluations idempotent
Summary:
This allows us to upsert the same evaluations in the asset daemon multiple times in the event that a tick fails

Sadly for this to work in sqlite it requires adding the evals one by one instead of in a batch. If you are using AMP on a large asset graph though, you should be using postgres or mysql, not sqlite.

Test Plan: BK