[ui] Disambiguate some SplitPanelContainer identifiers (#17272)
## Summary & Motivation

We're currently reusing the `explorer` identifier on SplitPanelContainer
for the asset graph explorer, job graph explorer, and resource explorer.

This means that they're all sharing the same `localStorage` key for
tracking the panel widths. There's no need to overload this value in
this way from what I can tell, so let's give them their own identifiers.

This might resolve an issue where the panel was unexpectedly hidden for
a user viewing the asset graph. That is, they may have collapsed the
panel in some other situation, but the collapsed state persisted here.

## How I Tested These Changes

Adjust split panel on explorer UIs, verify that the keys are separate in
localStorage.