[ui] Improve display of reported materialization events (#17246)
## Summary & Motivation

This PR makes a few small improvements to the experience around
"runless" reported materialization events:

- In the sidebar, the title of the section changes from `Materialization
in last run` to `Last reported materialization`. The "Run" row is
removed and a "Description" row is added. Note: I left "Materialization
system tags" and plots because I think those could be populated via the
API?

- In the asset list view and graph, links to the materialization (eg:
the timestamps) all go to the specific event entry on the asset events
tab rather than to a 404ing run page.

- Long descriptions truncate with a "Show more" in both the details page
and the asset graph sidebar.

After: (note the description is just a worst-case scenario of long copy
pasta)

![image](https://github.com/dagster-io/dagster/assets/1037212/204d52ea-0a34-49db-bd92-733124a4116c)


## How I Tested These Changes

Report materializations from the Materialize button on the Asset Details
page, see that things look a bit better. I also added a reported event
storybook to the sidebar stories.


![image](https://github.com/dagster-io/dagster/assets/1037212/0215af72-93a6-460c-b7a2-97485ca130bd)

Co-authored-by: bengotow <bgotow@elementl.com>