[dagster-embedded-elt] Fix bug when creating multiple Sling assets (#17283)
## Summary & Motivation

This adds a generated name to the @multi_asset decorator for build_asset, using the a combination of the source stream and target object which should be unique across all assets.

I took inspiration from the Airbyte/Fivetran packages which do something similar although they rely on a unique connector id which we do not have here. I can't think of a good
reason to have multiple assets with the same source/destination but if there was, then there would still be a collision here. Would appreciate feedback on whether it's a good
assumption.

## How I Tested These Changes

Added a test, failed, made changes, passed.