[ui] Autodetect links in context.log messages, render clickable links (#17236)
## Summary & Motivation

This is a fix for #15862. It's largely a port of the autolinker I wrote
for Mailspring
[here](https://github.com/Foundry376/Mailspring/blob/master/app/src/services/autolinker.ts),
which doesn't require external packages and has gotten a handful of bug
fixes over the years (MIT-licensed so it's safe to reference.)

The code uses a useEffect to look at the text node after it's rendered
in React and adds the `a` tags directly at the DOM level. All the links
open in new tabs and have rel=nofollow, noreferrer for safety's sake. It
uses useIdleCallback + a DOM treewalker so if the text is huge it
doesn't block the JS loop.

I also noticed that we render the full text of the logs into the log
rows, even if they're enormous. We only ever show 12 lines with a
conditional "Show More", so I added a sanity check that cuts the string
down to 15k characters. That should be enough to trigger "Show More",
even on a 4000px wide screen, and saves us from computing the bounding
box of such enormous blocks of text. Hopefully nobody has been logging
like this 😅


![image](https://github.com/dagster-io/dagster/assets/1037212/bcf74cb8-d6ae-445f-a0ba-63b360714eae)

## How I Tested These Changes

I created a test asset that logged both short and 500k+ text blocks
containing links.

---------

Co-authored-by: bengotow <bgotow@elementl.com>