Fix permisison checks when an asset key with no code location is under consideration (#17282)
Summary:
Asset keys that don't map to a particular code location (e.g. because
the asset was removed but it still has a backfill) were hitting a
KeyError in this permissions check when they try to index into the repo
handles in the asset graph. Instead, fall back to ignoring per-location
permissions and require deployment-wide permissions.

Test Plan: New test case

## Summary & Motivation

## How I Tested These Changes