[asset backfill] Fix 'all partitions' targeting (#17278)
Similar to https://github.com/dagster-io/dagster/issues/16952.

In certain situations when the `allPartitions = True` flag is applied to
an asset backfill, some downstream partitions hang and never get
executed.

For example, in the following situation:
- Asset A, Daily starting 2023
- Asset B, Daily starting 2022 (dependent on A) 

The following happens:
1. Backfill is launched with `allPartitions = True`, causing all
partitions of A and B are targeted
2. Backfill requests roots -- before the changes in this PR, that would
be all partitions of A
3. In subsequent iterations, partitions downstream of A partitions get
kicked off once A finishes executing. This would be partitions in B
after 2023
4. 2022 partitions in B hang indefinitely

Instead, we should target 2022 partitions of B as part of the 'request
roots' call.

This PR fixes this issue by updating the
`get_target_root_asset_partitions` call to also find partitions that do
not have parents in the target subset. Previously, it only found assets
that do not have parents in the target subset.