[pipes] Pipes opened payload processing (#17384)
## Summary & Motivation

This PR builds on the `opened` message processing in Pipes with the goal
of allowing message readers to perform processing dependent on
information sourced from the external process. The immediate impetus is
to allow the message reader to read DBFS logs whose location is only
known once a Databricks process has been started.

- `PipesMessageWriter` now has a `get_opened_payload` method. This is
called during the initialization routine in the external process, with
the returned "opened payload" passed back as parameters with the
`opened` message. Currently the opened payload has only one key,
"extras" , where an environment-specific message writer can put
arbitrary information. For the `PipesDbfsMessageWriter`, we use this for
the cluster log root path.
- I think we will want to add other keys to the opened payload in the
future, and we may want to have the `PipesDbfsContextLoader` contribute
to it as well.
- `received_any_message` on the `PipesMessageHandler` was changed to
`received_opened_message`. This is equivalent since the `opened` message
will _always_ be the first message received.
- `PipesMessageHandler` now takes `PipesMessageReader` in its
constructor instead of passing it to `handle_messages`. It calls an
`on_opened` callback on the `PipesMessageReader` when the opened payload
is received.
- `PipesStdioReader` has been renamed to `PipesLogReader`. This makes it
clear that the mechanism can read any log file, which may or may not
correspond to a stdio stream.
- `PipesMessageReader` now takes an `Optional[Sequence[PipesLogReader]]`
instead of dedicated slots for stdout/stderr readers.
- Because we now take a default empty list of `PipesLogReader`, the
`PipesNoOpStdioReader` is removed.
- The `start`/`stop`/`is_ready` interface of `PipesLogReader` was
retained, but is now called slightly differently in the
`PipesBlobStoreMessageReader`:
- `read_messages` now launches a separate "logs thread" from the main
message reading thread. This thread waits for the reception of the
opened payload and then launches the log readers. It then waits until
the external task is finished, at which point it spends up to a minute
waiting for logs to be written.
- The Databricks pipes integration and tests have been updated to use
this new system.

### Caveats

- I'm uncertain on the "opened payload" terminology`, but at least it's
pretty clear what it corresponds to.
- I have some reservations about the interaction of
`PipesMessageHandler` and `PipesMessageReader`. Currently the
implementation of `PipesMessageReader.on_opened` is left totally open
(no-op) by default, and this leads to a somewhat awkward situation in
`PipesBlobStoreMessageReader` where `on_opened` sets an attribute that
the `log_starter_thread` is polling for. I think this can probably be
improved.

## How I Tested These Changes

- Update databricks Pipes client tests