fix(dbt): allow empty exclude (#17412)
## Summary & Motivation
If an exclude statement comes up empty handed, still allow it. We should
only expect the select statement to have a non-empty result.

In the case that an exclude is empty, a set difference against it is
treated as if exclude was not populated to begin with.

## How I Tested These Changes
local, add a test