apply the asset checks loader on GrapheneAssetNode (#17400)
Pass in the loader object to GrapheneAssetNode so that we batch fetching
asset check definitions.

I think the optional loader pattern we've used with the other loaders is
a little sketchy. I'm open to pushback, but I think it's cleaner to
always pass in the loader. It means that we initialize it in some
unneeded places (which should be cheap, just adds code), but it means we
only have one code path when we read from the loader and we guarantee
that we'll use batching when we query for checks.

**Test plan**
added coverage for the resolver on AssetNode.

In terms of perf, the query

```gql
query {
  assetNodes {
  	assetChecks {
        name
    }
  }
}
```

used to take 3 minutes running on the big toys repo because it loaded
the asset graph for each of the ~2k assets. Now it takes 1 second.