[ui] Fix autolinker error in Safari (#17459)
## Summary & Motivation

It turns out that Safari does not support `window.requestIdleCallback`,
so extremely large logs in the log viewer cause the bottom portion of
the run log page to show an inline error box.

This PR makes the autolinker fall back to a "run for 500ms, with short
gaps to avoid blocking the event loop" strategy that just uses a
setTimeout. The goal at the end of the day is just to make sure
autolinking doesn't ever block the JS event loop and "lock" the page.

<img width="1420" alt="image"
src="https://github.com/dagster-io/dagster/assets/1037212/ca45cadc-c959-4b8d-9d0b-92ba400268c1">


## How I Tested These Changes

I wrote a new test case to verify that this fallback works, because
conveniently the jest env also does not have
`window.requestIdleCallback`. I also set the "work ms" to something very
low (15ms), and then did a perf trace on a 120,000 character string with
many links. The trace shows that it suspends and resumes and eventually
stops (good!) and scrolling to the bottom of the message indicates it
linkified the very last link (good!)


I also cruised around most of Dagster UI with Safari to make sure things
are generally looking good because it's been a while. Looking good!

Co-authored-by: bengotow <bgotow@elementl.com>