[ui] Fix crash when displaying a graph op with 4+ outputs or inputs (#17476)
## Summary & Motivation

This fixes a crash observed here
https://dagsterlabs.slack.com/archives/C058CBEL230/p1698075923055629

In our layout code, we "collapse" op inputs / outputs if there are >4.
(see `IO_THRESHOLD_FOR_MINI`). When this happens, we're not able to
"find" the dot that these input mappings are supposed to connect to in
the subgraph, which caused the graph to crash.

This fixes the IO layout lookup so that we can find the dot that the
inputs/outputs were "collapsed into". I also renamed several variables
in here to be more clear because it took me a while to remember what was
going on here.


![image](https://github.com/dagster-io/dagster/assets/1037212/edc4b824-d84a-4bf6-876f-a2dbbc79a659)


## How I Tested These Changes

I tested this by creating a new story for `OpGraph` that exhibits >4
outputs and >4 inputs. Hooray for stories 🙏

---------

Co-authored-by: bengotow <bgotow@elementl.com>