[AssetLiveDataProvider] Retry asset keys that fail to fetch after the poll interval. (#17393)
## Summary & Motivation

Currently if an asset key fails to fetch we never retry fetching it
again because it gets stuck in the requested state. This PR fixes that
and marks them as "fetched" when this happens. This has the consequence
of (1) letting us move on to the next set of asset keys which could
succeed and (2) it means we'll retry the failed asset keys again after
the polling interval threshold is crossed.

## How I Tested These Changes
jest. Locally threw an error and saw that we did end up retrying the
failed asset keys after the polling interval was met.